Refactor save system from PlayerPrefs to YG plugin saves

+++ Assets/Scripts/ShopSystem/IShopItem.cs
@@ -1,10 +1,8 @@
-using System.Numerics;
-
 public interface IShopItem
 {
     string ItemName { get; }
-    BigInteger ItemCost { get; }
-    BigInteger ItemBaseCost { get; }
+    ulong ItemCost { get; }
+    ulong ItemBaseCost { get; }
     int ItemLevel { get; }
     void BuyItem();
     void SaveBuyedItem();
